refactor specs for ves formatting
